requirements.txt
instagrapi==1.17.5
Pillow==9.5.0
python-dotenv==1.0.0
